:sparkles: Render the queries as templates
